Fix anti-forgery middleware failures
Because the `*anti-forgery-token*` is a binding, it does not play well
when being returned via a `def` - previously, I was defining the index
page HTML via a def. Once that was changed to a `defn` (or `defhtml`
which just wraps a function using hiccup's html macro), it started
properly displaying.

Need to learn more about bindings. I think the problem here was that the
`def` is compiled globally, whereas the `defn` gets called in the same
thread-local context where the binding gets properly bound.